Persistent changesgit status (kinda)
Changes persist across different tab sessions for a give page

A couple bugs still exist though. First, the first edit to a page does not persist, only the second onwards. This could be a logging issue or a syncronicity issue. Second, sticker edits affect all pages in storage. This is probably an edit parsing issue? Third, sticker edits can only be made when inspecting the popup. Otherwise, clicking the window causes the popup to go away and stop sticker functionality.